🛡️ Guardian: enforce C11 structure and union member constraints
This change hardens the semantic analysis of structure and union members
according to C11 6.7.2.1p3 and 6.7.2.1p18.

Key improvements:
- Rejects 'void', function types, and variably modified types (VLAs) as
  struct/union members.
- Improves Flexible Array Member (FAM) validation, ensuring it's only in
  structures and at the end.
- Refactors TypeRegistry layout computation to be robust against recursion
  and errors, preventing state leaks.
- Enhances diagnostics by re-mapping low-level layout errors to precise
  member-specific messages with correct spans.
- Added a comprehensive test suite in guardian_struct_member_constraints.rs.
- Fixed minor pre-existing clippy warnings in unrelated files.

+++ src/codegen/mir_gen_expression.rs
@@ -210,7 +210,7 @@ impl<'a> MirGen<'a> {
     fn visit_type_query(&mut self, ty: semantic::TypeRef, is_size: bool) -> Operand {
         let layout = self.registry.get_layout(ty);
         let val = if is_size { layout.size } else { layout.alignment };
-        self.create_size_t_operand(val as u64)
+        self.create_size_t_operand(val)
     }
 
     fn visit_generic_selection(
